requirements.txt
depthai
opencv-python
pyyaml
